refactor #2 - rename package and add some utils
- 패키지 이름을 domain에서 model로 변경합니다.
- logging 편의성을 위해 @ToString을 적용합니다.
- 테스트 편의성을 위해 @Builder를 적용합니다.
